requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
reportlab==4.0.7
gunicorn==21.2.0
openai==1.12.0
anthropic==0.18.0
google-generativeai==0.3.2
